Seção 9 - DTOs e Data Annotation

+++ BackEnd/ProEventosAPI/ProEventosAPI/Controllers/EventosController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using ProEventos.Application.Dtos;
 using ProEventos.Application.IRepositoryApplication;
 using ProEventos.Domain;
 using System;
@@ -23,8 +24,7 @@ namespace ProEventosAPI.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) 
-                    return NotFound("Eventos não encontrado.");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -41,8 +41,7 @@ namespace ProEventosAPI.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null)
-                    return NotFound("Evento por Id não encontrado.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -59,8 +58,7 @@ namespace ProEventosAPI.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (eventos == null)
-                    return NotFound("Eventos por tema não encontrado.");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -72,13 +70,12 @@ namespace ProEventosAPI.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.AddEvento(model);
-                if (evento == null)
-                    return BadRequest("Erro ao tentar adicionar evento.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -90,13 +87,12 @@ namespace ProEventosAPI.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
-                if (evento == null)
-                    return BadRequest("Erro ao tentar atualizar evento.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -112,9 +108,12 @@ namespace ProEventosAPI.Controllers
         {
             try
             {
-                return await _eventoService.DeleteEvento(id) 
+                var evento = await _eventoService.GetEventoByIdAsync(id);
+                if(evento == null) return NoContent();
+
+                return await _eventoService.DeleteEvento(id)
                        ? Ok("Evento deletado.") 
-                       : BadRequest("Erro ao tentar deletar evento.");
+                       : throw new Exception("Ocorreu erro não específico ao tentar deletar Evento.");
                 
             }
             catch (Exception ex)
